Moved all the website stuff from the repository
Previously, we built this course from the repository for DataCamp. However, some of those files shouldn't be seen by students. So, we wanted to make the repository private. But, that would make the gh-pages branch stop working. So, we created this new repository just for the website and moved over all the files from the old repository.